[DX] Move LocaleRepositoryInterface from bundle to component
