+++ spec/Pim/Bundle/BaseConnectorBundle/Processor/CsvSerializer/HeterogeneousProcessorSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\BaseConnectorBundle\Processor\CsvSerializer;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 use Prophecy\Argument;
 use Symfony\Component\Serializer\SerializerInterface;
