+++ spec/Pim/Bundle/CatalogBundle/Validator/AttributeValidatorHelperSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\CatalogBundle\Validator;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Validator\AttributeValidatorHelper;
 
 class AttributeValidatorHelperSpec extends ObjectBehavior
