+++ spec/Pim/Component/Connector/ArrayConverter/Flat/AttributeOptionStandardConverterSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Component\Connector\ArrayConverter\Flat;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 
 class AttributeOptionStandardConverterSpec extends ObjectBehavior
 {
