+++ spec/Pim/Component/Connector/ArrayConverter/Flat/Product/AttributeColumnInfoExtractorSpec.php
@@ -8,7 +8,7 @@ use Pim\Component\Catalog\Model\ChannelInterface;
 use Pim\Component\Catalog\Model\LocaleInterface;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 
 class AttributeColumnInfoExtractorSpec extends ObjectBehavior
 {
