+++ src/Pim/Bundle/BaseConnectorBundle/Processor/CsvSerializer/ProductProcessor.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\BaseConnectorBundle\Processor\CsvSerializer;
 
 use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Constraints as Assert;
