+++ src/Pim/Bundle/CatalogBundle/Validator/AttributeValidatorHelper.php
@@ -3,7 +3,7 @@
 namespace Pim\Bundle\CatalogBundle\Validator;
 
 use Pim\Bundle\CatalogBundle\Repository\ChannelRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Repository\LocaleRepositoryInterface;
+use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
 use Pim\Component\Catalog\Model\AttributeInterface;
 
 /**
